Initial commit! Just getting started.
I am currently just working on the HTML by hand. I was trying to make
something true to the era, but my dear god is it a PITA to do without
CSS or anything. I am still working out of a single page as I do not
have so much content that I need a static site generator, but I might in
the future move to something like that just for my sanity's sake so I
can pump out more content without having to go through too much
headache.